reorg package and remove duo ModuleDescriptor
